% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquare_functions.R
\name{rsquare_explore}
\alias{rsquare_explore}
\title{rsquare_explore function}
\usage{
rsquare_explore(latlon, token, date, radius, simplify = TRUE)
}
\arguments{
\item{simplify}{This returns a tidy dataframe, defaults to TRUE}
}
\description{
This function accesses the foursquare explore api and returns a json response and/or a dataframe
}
\examples{
rsquare_explore("40.75,-73.98","key/token","20161028",2000, simplify = TRUE)
}
\keyword{rsquare}
