% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquare_functions.R
\name{rsquare_search}
\alias{rsquare_search}
\title{rsquare_search function}
\usage{
rsquare_search(latlon, token, date, radius, simplify = TRUE)
}
\arguments{
\item{simplify}{This returns a tidy dataframe, defaults to TRUE, in BETA, parsing list fix in development.}
}
\description{
This function accesses the foursquare search venues api and returns a json response and/or a dataframe
}
\examples{
rsquare_search("40.75,-73.98","key/token","20161028",2000, simplify = TRUE)
}
\keyword{rsquare}
